Refactor: API 호출 로직 함수 분리
