+++ src/components/common/PlaylistAction.tsx
@@ -3,8 +3,8 @@ import { useEffect, useState } from "react";
 import BookmarkIcon from "@/assets/icons/bookmark.svg?react";
 import CommentIcon from "@/assets/icons/comment.svg?react";
 import HeartIcon from "@/assets/icons/heart.svg?react";
-import { usePlaylistDetail } from "@/hooks/usePlaylistDetail";
-import { usePlaylistActionInfo } from "@/hooks/usePlaylistActionInfo";
+import { usePlaylistDetail } from "@/hooks/queries/usePlaylistDetail";
+import { usePlaylistActionInfo } from "@/hooks/queries/usePlaylistActionInfo";
 import axiosInstance from "@/services/axios/axiosInstance";
 import useUserStore from "@/store/useUserStore";
 import { useQueryClient } from "@tanstack/react-query";
