+++ src/layout/Header.tsx
@@ -6,7 +6,7 @@ import Search from "@/assets/icons/search.svg?react";
 import Logo from "@/assets/imgs/logo.svg?react";
 import OverflowMenu from "@/components/common/OverflowMenu";
 import SearchBar from "@/components/common/SearchBar";
-import { usePlaylistDetail } from "@/hooks/usePlaylistDetail";
+import { usePlaylistDetail } from "@/hooks/queries/usePlaylistDetail";
 import axiosInstance from "@/services/axios/axiosInstance";
 import supabase from "@/services/supabase/supabaseClient";
 import useUserStore from "@/store/useUserStore";
